1.22_4

+++ pkg/ldacs_core/model/account_gs.go
@@ -8,7 +8,7 @@ import (
 // 地面站 结构体  AccountGs
 type AccountGs struct {
 	global.PREFIX_MODEL
-	GsSac       int     `json:"gs_sac" form:"gs_sac" gorm:"column:gs_sac;comment:;"`                   //地面站SAC
+	GsSac       uint64  `json:"gs_sac" form:"gs_sac" gorm:"column:gs_sac;comment:;"`                   //地面站SAC
 	LatitudeN   float64 `json:"latitude_n" form:"latitude_n" gorm:"column:latitude_n;comment:;"`       //北纬
 	LongtitudeE float64 `json:"longtitude_e" form:"longtitude_e" gorm:"column:longtitude_e;comment:;"` //东经
 }
